% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\docType{data}
\name{bf}
\alias{bf}
\title{Multiple Sources of Attentional Dysfunction in Adults With Tourette’s
Syndrome}
\format{
A data frame with 51 rows and 2 columns:

\describe{
\item{\code{accuracy}}{(numeric) Participant's accuracy in the attentional
task}
\item{\code{group}}{(factor) Participant's group membership (TS patient,
ADHD patient, or control).}
}
}
\source{
\code{{BFpack}} package
Last retrieved from CRAN: 2024-10-12
}
\usage{
bf
}
\description{
Data from a psychological study comparing attentional performances of
Tourette’s syndrome (TS) patients, ADHD patients, and controls. These data
were simulated using the sufficient statistics from Silverstein et al.
(1995).
}
\references{
Silverstein, S. M., Como, P. G., Palumbo, D. R., West, L. L.,
& Osborn, L. M. (1995). Multiple sources of attentional dysfunction in
adults with Tourette’s syndrome: Comparison with attention deficit-
hyperactivity disorder. Neuropsychology, 9(2), 157–164.
https://doi.org/10.1037/0894-4105.9.2.157
}
\keyword{datasets}
