% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R
\name{compare2}
\alias{compare2}
\title{Print two columns of interest from the data frame in the console.}
\usage{
compare2(data, column1, column2)
}
\arguments{
\item{data}{The data frame in question.}

\item{column1}{The first column to compare.}

\item{column2}{The second column to compare.}
}
\value{
The two selected columns printed in the console.
}
\description{
\code{compare2()} is a function that prints two columns of interest from a data
frame side by side, for a quick visual row-wise comparison.
}
\examples{
compare2(mtcars, 'cyl', 'mpg')

}
