% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecone.R
\docType{data}
\name{doublecone}
\alias{doublecone}
\title{Sub-clinical ADHD behaviors and classroom functioning in schoolage children}
\format{
A data frame with 686 observations on the following 4 variables:

\describe{
\item{sex}{(factor) 1=boy; 2=girl}
\item{ethn}{(factor) 1=Colombian, 2=African American, 3=Hispanic American,
5=European American}
\item{hypb}{(numeric) Classroom hyperactive behaviour level.}
\item{fcn}{(numreric) A measure of social and academic functioning.}
}
}
\source{
\code{{DoubleCone}} package
Last retrieved from CRAN: 2024-10-12
}
\usage{
doublecone
}
\description{
Observations on children aged 9-11 in classroom settings, for a study on the
effects of sub-clinical hyperactive and inattentive behaviors on social and
academic functioning.
}
\references{
Brewis, A., Schmidt, K. L., & Meyer, M. (2000). ADHD‐Type
Behavior and Harmful Dysfunction in Childhood: A Cross‐Cultural Model.
American Anthropologist, 102(4), 823–828.
https://doi.org/10.1525/aa.2000.102.4.823
}
\keyword{datasets}
