% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_val.R
\name{max_val}
\alias{max_val}
\title{Get the maximum value in each numeric column of a data frame.}
\usage{
max_val(data)
}
\arguments{
\item{data}{The data frame in question.}
}
\value{
The maximum value in each of the numeric columns in the specified
data frame (provided there are numeric columns in the data frame).
}
\description{
\code{max_val()} is a function that returns the maximum value in each of the
numeric columns of a data frame.
}
\examples{
max_val(mtcars)
max_val(ChickWeight)

}
