% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_val.R
\name{min_val}
\alias{min_val}
\title{Get the minimum value in each numeric column of a data frame}
\usage{
min_val(data)
}
\arguments{
\item{data}{The data frame in question.}
}
\value{
The minimum value in each of the numeric columns in the specified
data frame (provided there are numeric columns in the data frame).
}
\description{
min_val() checks for any columns in a data frame that contain numeric values,
then returns the lowest value in each of those numeric columns.
}
\examples{
min_val(mtcars)
min_val(ChickWeight)
}
