% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcollectadhd-package.R
\name{rcollectadhd}
\alias{rcollectadhd}
\title{rcollectadhd: Tools to perform common simulation tasks}
\description{
This package includes seven data frames which have been sourced from other
packages on CRAN or from publications on Kaggle. The package also includes
four very simple functions for obtaining basic statistics about a data frame.
}
