% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R
\name{compare2}
\alias{compare2}
\title{Print two columns in the console}
\usage{
compare2(data, column1, column2)
}
\arguments{
\item{data}{The data frame in question}

\item{column1}{The first column you want to compare.}

\item{column2}{The second column you want to compare.}
}
\value{
The two selected columns printed in the console.
}
\description{
The point of compare2() is to reproduce two columns of interest side by side
in the console for a quick scan comparison row by row.
}
\examples{
compare2(mtcars, 'cyl', 'mpg')
compare2(ChickWeight, 'Chick', 'Time')
}
