% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecone.R
\docType{data}
\name{doublecone}
\alias{doublecone}
\title{Sub-clinical ADHD behaviors and classroom functioning in schoolage children}
\format{
A data frame with 686 observations on the following 4 variables

\describe{
\item{sex}{1=boy; 2=girl}
\item{ethn}{1=boy; 2=girl}
\item{hypb}{1=boy; 2=girl}
\item{fcn}{1=boy; 2=girl}
}
}
\source{
\code{{DoubleCone}} package
}
\usage{
doublecone
}
\description{
Observations on children aged 9-11 in classroom settings, for a study on the
effects of sub-clinical hyperactive and inattentive behaviors on social and
academic functioning.
}
\references{
Brewis, A., Schmidt, K. L., & Meyer, M. (2000). ADHD‐Type
Behavior and Harmful Dysfunction in Childhood: A Cross‐Cultural Model.
American Anthropologist, 102(4), 823–828.
https://doi.org/10.1525/aa.2000.102.4.823
Last retrieved from CRAN: 14/10/24
}
\keyword{datasets}
