% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_val.R
\name{min_val}
\alias{min_val}
\title{Get the minimum value in each numeric column of a data frame}
\usage{
min_val(data)
}
\arguments{
\item{data}{The data frame in question.}
}
\value{
The minimum value in each of the numeric columns in the specified
data frame (provided there are numeric columns in the data frame).
}
\description{
\code{min_val()} is a function that returns the minimum value in each of the
numeric columns of a data frame.
}
\examples{
min_val(mtcars)
min_val(ChickWeight)

}
